landmark_placement_mvlm/requirements.txt
absl-py==0.9.0
alabaster==0.7.12
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.3
Babel==2.8.0
backcall==0.2.0
bleach==3.1.5
blinker==1.4
brotlipy==0.7.0
cachetools==4.1.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1594141588948/work
cryptography==2.9.2
cycler==0.10.0
cytoolz==0.10.1
dask @ file:///tmp/build/80754af9/dask-core_1594156306305/work
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
docutils==0.16
entrypoints==0.3
flake8==3.8.3
future==0.18.2
google-auth @ file:///tmp/build/80754af9/google-auth_1594357566944/work
google-auth-oauthlib==0.4.1
grpcio==1.27.2
helpdev==0.7.1
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
imagesize==1.2.0
importlib-metadata==1.7.0
intervaltree==3.0.2
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.1
jeepney==0.4.3
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.5
jupyter-core==4.6.3
keyring==21.2.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1592846017414/work
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.0
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.7
networkx @ file:///tmp/build/80754af9/networkx_1594377231366/work
numpy==1.18.5
numpydoc==1.1.0
oauthlib==3.1.0
olefile==0.46
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1594307325547/work
pluggy==0.13.1
prompt-toolkit==3.0.5
protobuf==3.12.3
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydocstyle==5.0.2
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.8.0
PyQtWebEngine==5.12.1
pyrsistent==0.16.0
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.34.1
pytz==2020.1
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole==4.7.5
QtPy==1.9.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rope==0.17.0
rsa==4.0
scikit-image==0.16.2
scipy @ file:///tmp/build/80754af9/scipy_1592930510235/work
SecretStorage==3.1.2
six==1.15.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
Sphinx==3.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.4
spyder-kernels==1.9.2
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0
testpath==0.4.4
toml==0.10.1
toolz==0.10.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
ujson==1.35
urllib3==1.25.9
vtk==8.2.0
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
wurlitzer==2.0.1
yapf==0.30.0
zipp==3.1.0
